myBlog/djangoProject/requirements.txt
django
Pillow
python-slugify